Update database model (#23)
* Add layers to the database model to clarify different parts of the schema

* Revert column renaming

Rename jatehuoltomaarays back to velvoitemalli. This column
was renamed to jatehuoltomaarays for clarity for demo sessions.
It had no yet migrations etc.

* Remove regulation check functions

These are client/implementation specific so should not be in core repo.

* Fade in kiinteisto

* Change relationships to viranomaispaatos to manual

PgModeler had a "bug" that automatic relationships even with sql-off
came to diff files. Overcome this by changing to manually created
constraints.

* Changes to model to make diff cleaner

Co-authored-by: Lauri Kajan <lauri.kajan@gispo.fi>