Finalized setting rakennuksen_omistajat omistuksen_loppupvm.
The finalized function sets omistuksen_loppupvm as follows:
Checks for existing entries that are present in the imported dvv, and selects the omistuksen_alkupvm - 1 day, if the existing entry's alkupvm is greater. If multiple existing entries are found, picks the entry with latest (greatest) omistuksen_alkupvm. If no existing entry is found, or  exsisting entry's omistuksen_alkupvm is not greater, then poimintapvm is used.