Added report function for targets obligations (kohteiden velvoitteet).
The function returns a table of targets obligations that are fulfilled. For each type of obligation, the description of the fulfilled obligation is added. If a target does not fulfill any obligations for the type, the field is left empty.

If multiple obligation checks overlaps with the given "tarkistuspvm", use the one with later upper value in the daterange.